ibssa: Define directory structure
Refine the source tree directory structure.  The structure
separates the SSA code base into the 4 identified node types.
The proposed structure is:

include - common header files
include/infininband - installed header files
master
	The master SSA connects directly to the SM as a
	plugin.
distrib
	The distribution layer is responsible for quickly and
	efficiently exchanging data up and down the SSA
	network of nodes.
access
	End users connect into the SSA network at the access
	layer.  The access layer converts raw data into a format
	which can be consumed
acm
	This is the client code that connects into the SSA
	tree to extract the data necessary to establish
	communication with other clients.

The names for the distribution and access layers are based on
Cisco's 3 layer networking model.  (Cisco refers to the top
layer as the core layer, but that was left as master.)

Signed-off-by: Sean Hefty <sean.hefty@intel.com>