ibssa: Create common calls to open devices/ports
Generalize ibacm code to open all devices/ports.  The
SSA common code will open/close all devices.  To support
the different user's of the common code, the size of the
device and port structures that are allocated are specified
by the caller.

The intent is that ssa_device and ssa_port will be embedded
inside user specific structures

Signed-off-by: Sean Hefty <sean.hefty@intel.com>