Move SSA message protocol to infiniband/ssa.h
ibssa_control.h contained a very early attempt at
defining what the inter-node protocol would be for SSA
nodes.  Move this file from include/ibssa_control.h
to include/infiniband/ssa.h.  The rename helps to
avoid confusion with ssa_ctrl.h, which defines
intra-node SSA messages.  The move is done with the
assumption that external protocol files will be
installed as part of an installation package.

The messages defined in the file are cleaned up.
Most of the file included structures and definitions
defined in acm.h.  Rather than duplicating those
definitions, we can use a modified version of acm.h
when we're ready for that part of the implementation.

The base message structure is changed to remove the
IB MAD header.  An SSA specific header is defined
instead (though clearly derived from the MAD header).
To help us isolate disjoint message protocols (e.g.
ACM from the SSA DB), we use define an SSA 'class'
for the message types.  Two classes are defined:
a control class and a DB class.

A third class is included as a possibility, but
commented out -- an encapsulated MAD class.  It can
either be added when needed or removed if we decide
not to support standard MAD operations over SSA
connections.  It was added in response to Ira's comment
in the ibssa_control file about supporting both
standard SA queries and optimized (i.e. acm) queries.

Each class can define their own operations and structures
without worry of conflicts.  The only control message
which is defined is a 'sync' message, which is intended
to be used as part of the connection process.  A parent
will send a sync message after establishing a connection
with a child to inform the child that the connection
has completed successfully.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>