Remove top-level build files
Building from the root of the source tree is not usable.
The build packages are available under the acm and plugin
directories, with the plugin autotool files being derived
versions from the source tree root.

Remove the autotool files from the top of the source tree
and move the spec file into the plugin directory.  The spec
file was not changed, just moved.  It has not been tested.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>