ssa: Separate ID into fields
Define an ID with database, table, and field components.  If
an ID is set to DB_ID_DEFS, document that definitions should
be returned, rather than data.

With this change, it becomes easier to request all data from
a database.  A response may contain multiple datasets.  To
simplify the responder, allow a response to be variable size,
where the full size of the response is not known up front.  This
is handled by adding an END flag to the SSA message header.  The
operation would thus be:

Request asks for all data from a database.
Responder sends ssa_msg_hdr with len = -1
Responder sends datasets for all data in the database
Responder sends ssa_msg_hdr with END flag set

Signed-off-by: Sean Hefty <sean.hefty@intel.com>