Initial implementation of setting parent for prototype
When join (MemberRecord) is received, in addition to being accepted,
this now initiates an SA query for a path between the joined
node and the core (which is the parent in the prototype).
When the PathRecord is returned from the SA, the core
then issues a set parent (InfoRecord) which includes the
path record for a subsequent connection setup (via rsockets).
The joined node also acknowledges the set parent with a
get response.

A few notes (on some TODOs):
1. There is currently an artificial delay before initial path
query. This needs fixing.
2. Tree calculation is just noted and needs to be added to framework.
3. A per client state machine (at the core) needs to be documented
(designed) and implemented.
4. Add component masks to umad_sa.h and change sa_init_path_query
comp mask setting accordingly.
These will be subsequent patches.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>