ssa.c: In ssa_upstream_update_conn, add temporary delay before first query
Delay is currently 10 msec. This is dependent on time to calculate
half world PRs (and some thread context switches).

This is a workaround to be replaced by real solution later.

First query from ACM after connection established arrived too soon
at the access layer - prior to access layer doing it's work to
compute the PRDB for that ACM client.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>