core.c: Enhance core_build_tree to support single core <-> access <-> ACM
Only partially tested because only 2 nodes available.
Portion untested is whether the PR query for the ACM works and
whether the ACM is properly parented to the access and
is able to establish the connection. If it doesn't work, it's
a minor amount of debug when we're ready for this.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>