ssa.c: Add access layer DB update message queue
Rather than the access layer send PRDB update
message directly to the downstream thread,
a queue is added for holding PRDB updates. It
is populated by access layer and later processed
by access_prdb_handler thread, that sends the
updates to downstream thread.

This is in preparation for the multithreaded
access layer.

Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>