acm/acme.c: Improve PR verification
Currently if running 'ib_acme' with '-v' flag,
the only verification that is done, is checking
if the same path record exists in SA.

The verification is improved by adding comparison
of path record parameters (currently only 'mtu' and
'rate' are checked, but more parameters will be added
in future).

This verification improvement is needed for better and
more precise testings on Kodiak cluster.

Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>