ssa.c: In ssa_access_handler, update access map on (ACM) client reconnection
Handle reconnection of ACM client in following scenario:
ibacm
...
pkill -9 ibacm
ibacm

Also, better handling of subsequent reset of "uninvolved" port in
SSA distribution tree
Same scenario as above with ibportstate reset of "uninvolved" port

When the downstream thread accepts connection for PRDB, it sends
SSA_CONN_DONE message to access thread. The client (ACM) GID is looked up
in the access map. When found, if the rsocket in the connection done message
is different from that in the map, it (the consumer entry in the map)
is updated to the new connection information (both rsock and LID).

Signed-off-by: Hal Rosenstock <hal@mellanox.com>