Add "fake" (ACM) clients to access
"Fake" clients are used for path record computation testing. The "fake" client
has a gid like real ACM host connected to the subnet.
For the client, access performs the path record calculation, but doesn't
send resultant DB calculated.

A new build configuration parameter is added:
 	--enable-fake-acm
The default value is "disable".

A new parameter is also added to access configuration:
 	fake_acm_num
It's a numeric parameter and controls max. number of "fake" clients
added to a service.
If it's equal 0, no "fake" clients are added.
If it's less than 0, all hosts from subnet will be added as "fake" clients.

Signed-off-by: Sasha Kotchubievsky <sashakot@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>