ssa_path_record.c: Fix crash in case of port not found
find_linked_port and find_port functions changes current
port. Current port lid and number are used as input parameters
for the functions. If the function can't find a new port, it
returns NULL. The same pointer to port is used for formatting
an error message. It crashes because the pointer was changed
to NULL. Solution is to use temporary variables.

Signed-off-by: Sasha Kotchubievsky <sashakot@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>