ssa.c: Use proper variable for accessing downstream thread fds array
Currently in order to access '**fds' array in ssa_check_listen_events()
routine, '*pfd' pointer, which is method argument, is used.

Instead of that '*pfd2' pointer, that is localy defined,
will be used.

There is no issue with current functionality, but it seems
like a wrong and confusing way that things are done.

Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>