ssa.c: In ssa_check_listen_events, fix fd index of downstream connection that is being closed
In the case of downstream node that reconnects, the old connection
is closed. Beside closing the proper rsocket that was used
for old connection, proper fd in downstream thread should be
cleaned as well.

Currently, the wrong fd is closed, because it is assumed that '**fds'
array have the same index as 'fd_to_conn[]' for the same connection,
which is rsock, and that is not true.

Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>