ssa.c: In ssa_downstream_svc_server, close connection if update pending or waiting
While DB update is pending or waiting, incoming connections cannot
be accepted.

Unfortunately, in rsockets, connection must be accepted before torn down in
case of update is pending or waiting so that is what this change does.

This relies on downstream client reconnecting some time later.
This part is not currently implemented (can only kill and restart
downstream client to "simulate" this behavior).

Signed-off-by: Hal Rosenstock <hal@mellanox.com>