plugin/ssa_extract.c: Eliminate full LFT extract dependency on 'first' flag
Currently full LFT tables extraction is conditionalized
on 'first' flag, that is raised after first SUBNET UP event.
This flag is also used in other SSA mechanisms, such as
distribution tree generation and thus it might interrupt
LFT tables extraction.

In order to avoid that, a simple check for already existing
LFT tables is performed, and if not - full LFT tables extraction
is forced.

Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>