ssa.c: Change UPSTR and DNSTR thread names in ssa_start_svc
so that they fit in 16 character limitation for thread name
due to pthread_setname_np

UPSTR now just UP and DNSTR now just DN to save 3 characters.

svc name was being truncated as follows:
Rather than mlx4_0:1:1, only mlx4_0:1: was in thread name in log

This is a workaround as arbitrary IB device names can be longer.
We are most interested in mlx4, mlx5, and qib.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>