core.c: Preserve SMDB address in case of no SMDB change
Currently, after each SUBNET UP event, previous version
of SMDB is destroyed and a new one is generated. As a
consequence of this, the SMDB pointer is changed even if
the data remained the same.

Such scenarios break running SSA plugin as combined
node (core + access layers) where the SMDB pointer
is shared between core and access layers.

Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>