ssa.c: Fix SSA_PR_SUCCESS handling in ssa_calculate_prdb
This has been broken since the signature for ssa_pr_compute_half_world
was changed. It results in successful PRDB calculation logging
a failure message and improper dumping depending on the option settings.
The above is not critical to proper operation but is confusing
to someone (like me) reading and debugging via the log.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>