ssa.c: In ssa_access_handler, access_context.smdb NULL is not an error
on receipt of SSA_CONN_DONE message

It is possible that the upstream is not connected or the SMDB has not
yet been received when the downstream connection occurs. So this is
"normal" and not an error.

In this case of access_context.smdb NULL, the access consumer needs
to be added to the core member map so that later when the SMDB DB
update occurs (after the SMDB SSA DB is received on the upstream
connection), the PRDB will be recalculated as well as the updated
epoch being riowritten into the consumer/ACM causing it to pull
it's updated PRDB.

This was tested by inserting immediate rejoin (ssa_svc_join(svc))
in ssa_upstream_mad when status is rejected. This was a temporary
hack for testing. The real mechanism of implementing a timed retry
with exponential backoff up to some max needs to be implemented.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>