core.c: Fix melded topology flow (subnet up event handling in core_report)
When running  CORE layer with  melded  topology, a different
plugin callback (core_report) is registered in OpenSM.
This callback responds to different SM events only
printing messages to the log (instead of sending
messages to extract thread as is done in the "regular" flow).

The patch below eliminated the need in 'subnet_up_delay'
parameter, but also moved some of the initial distribution
tree functionality to extract thread, ('first' flag reset
was moved from core_process_join() to extract thread).

As a consequence of that, when running CORE layer with
melded topology ('--enable-simulated-smdb'), no distribution
tree is generated, and thus there are no SMDB updates
between CORE and downstream nodes.

Patch that introduced the problem:

commit ab4c402f6b50093983b658e0392cf283107f419f
Author: Sasha Kotchubievsky <sashakot@mellanox.com>
Date:   Sun Sep 7 15:25:26 2014 +0300

    core.c: Remove subnet_up_delay parameter

    If subnet is not yet up, received SSA membership requests are
    put into an "orphan" list. The distribution tree is built
    after receiving "SUBNET UP" event,

Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>