ssa.c: In ssa_upstream_handler, workaround possible cause of async rconnect issue
It has been observed when running valgrind that the ACM upstream
connection request is issued, accepted at upstream side, but is
ACM never detects that is was completed.

It is theorized that this is due to needing to rely on errno
which could be set to errors other than EINPROGRESS.

With this scheme implemented by rsockets, can't have 2 upstream
connections in progress at same time.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>