ssa.c: Reconnect in the case of an rsocket error
This feature is part of SSA resiliency. It addresses non-core node
resiliency. The feature allows rsocket connection to be reestablished
with an upstream node.

The node recognizes connection failure using rsocket keepalives.
If node reaches the max. number of reconnection attempts, it gives up
and tries rejoin to the distribution tree.

Reconnection policy: This patch implements a constant reconnection policy.
A node reconnects to upstream node with a fixed/constant timeout
in between each reconnection attempt and limits of max. number of
attempts before giving up.

Signed-off-by: Sasha Kotchubievsky <sashakot@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>