core.c: Invoke calloc in with num_svcs rather than 1 as number of members
The calculation is correct. There is no real problem here.
extract_data.svcs is array of pointers. sizeof(extract_data.svcs) ==
sizeof(*extract_data.svcs). It's the reason that calculation is correct.
But from pure code style, we should pass to calloc size of element in
array and not a size of pointer to array.

It causes alert in static code analysis (coverity).

Signed-off-by: Sasha Kotchubievsky <sashakot@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>