ssa.c: In ssa_calculate_prdb, return original/current PRDB when ret != 1
Either new PRDB is same as/equal to current PRDB or
invalid PRDB structure is detected during compare.

It is necessary to return original/current PRDB so update
is generated in case of restarts of access and ACM.

If there is some optimization for this, it can occur
in downstream thread (not done currently).

Signed-off-by: Hal Rosenstock <hal@mellanox.com>