ssa_release_notes.txt: Updated librdmacm information and date
Updated due to recent changes/incorporation of critical
SSA patches in librdmacm. In at least one case, upstream
patch was not identical to one initially developed and in
use for January release of an SSA capable librdmacm package.
There were also additional patches incorporated beyond
that release.

There is no current released package for librdmacm that
is "up to date" for SSA.

Signed-off-by: Hal Rosenstock <hal@mellanox.com>