libadmin.c: Refactoring - Implement polymorphic message processing
The main reason for the refactoring is code reuse. Each command has
own set of parameters, type of a request message, response handling
and help string. But workflow is similar for different commands.

A new command should be added as a set of callbacks implementing
different aspects of command processing. No "if"s and "switch"s
in common routines.

This commit adds a "counter" command too. No real implementation
is included into the commit. It just illustrates how common
data structures could be used for two different commands: "ping"
and "counter".

Notes:
1. Command's properties are placed into 3 structures:
	- struct cmd_struct: The struct includes name, id, type.
	  The properties are used in both layers: application and
          lib, so the struct is moved from admin.c to libadmin.h.
	- struct cmd_struct_impl: The struct includes functions used
	  for command processing.
	- struct admin_command: The struct is an instance of the commnad
	  It includes command parameters and runtime information.
struct cmd_struct and struct cmd_struct_impl are "class" of a command
and include information relevant for all commands of the same type.
struct  admin_command is a "object". The object is created using
init function based on parameters passed to command line. It should
be destroyed using "destroy" function.
struct cmd_struct and struct cmd_struct_impl could be united in
the future. The reason for two structures is splitting SSA Admin
to command line application layer and a library that the appl use.

2. admin_exec implements polymorphic command processing. The flow
   includes following steps:
	- parsing command line and creating an instance of a command.
        - preparing an request message
        - sending the request to a node
        - handing response
        - destroying the command's instance.

3. struct cmd_help includes callbacks and an information relevant
   for preparing command's help. The struct is a part of
   struct cmd_struct_impl.
   Help's callbacks now receives FILE* as input and prints an
   information to a stream rather than returning fixed strings.

4. struct admin_context is used for passing an environment
   parameters to different stages of command processing.
   Right now, it includes only execution start and finish timestamps.

5. Command's implementation could use some default implementation.
   For example, if a command (for example ping) hasn't command line
   parameters it could use default_init.

6. "commands" array is renamed to "admin_cmds". The array
    now used in two compilation modules and should be "extern".
    "commands" name is not fit for global variable used in more than
    one compilation module.

Signed-off-by: Sasha Kotchubievsky <sashakot@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>