acme.c: Add mode option to ib_acme
ACM mode option is added to ib_acme in order to
generate mode specific options file. If no mode is specified,
it assumes default that SSA mode is to be used. In order
to produce options file for ACM mode, the following should
be executed: 'ib_acme -O -M acm'

Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>