ssa.c: Fix crash when an empty data table is received
In case of SSA DB with empty data tables received, all
proper ssa_db pointers will be assigned with the same
memory address and that will eventually lead to double
free call using the same address.

In order to avoid this, an additional check for received
buffer is added and if it is the same as message header,
no data table assignment will be done.

Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>