acm.c: Run SSA service on first active port that is found
Instead of running SSA service on first port of the first device
detected, we look for the first IB port in ACTIVE state, if no
such port found, we use the first port in INIT or ARMED states.

Only if all ports are DOWN, the first port is being used as it
was done before.

Signed-off-by: Ilya Nelkenbaum <ilyan@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>