libadmin.c: Remove separate printf for GID prefix in node_info_command_output
According to format option, nodeinfo command could print nothing
on the screen. For example, if the destination SSA node is ACM and
user asks to print downstream connection, nodeinfo's output is
empty. Empty output along with recursive mode (in the mode GID
address is printed as prefix) produces non valid string on the screen.

Signed-off-by: Sasha Kotchubievsky <sashakot@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>