Add "disconnect" command to ssadmin
The command breaks the data connection in SSA node.
After the disconnect, SSA nodes restore connections using
reconnect or rejoin mechanisms.

The command receives address of a connection as a command parameter.
The address could be either the GID or LID of upstream or downstream
connection.

"--addr_type" command option controls the type of the address.
Possible values:
	lid
	gid

Examples:
	ssadmin -l 2808  disconnect --addr_type=lid 6
In the example above, 2808 is a LID of destination node, 6 is a LID
of connection that should be disconnected.

	ssadmin -l 2808  disconnect --addr_type=gid fe80::8:f104:39a:169
In the example above, fe80::8:f104:39a:169 is a GID of connection
that should be broken.

The "disconnect" command is advanced functionality that will
be used in automatic test. It will be hidden from end-user using
conditional compilation and not mentioned in ssadmin man pages.

Signed-off-by: Sasha Kotchubievsky <sashakot@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>