ssa.c: Avoid double deletion of PRDB in ssa_calculate_prdb
Previously, if in function ssa_calculate_prdb the new prdb was computed
but the attempt to copy it failed - the PRDB pointed to by prdb_current
was destroyed (because of the 'prdb_current = prdb' assignment).
This was followed by another deletion at the next SMDB update.
This patch avoids that scenario.

Signed-off-by: Roy Menczer <roym@mellanox.com>
Signed-off-by: Hal Rosenstock <hal@mellanox.com>